# feat: Implement Phase 1 & 2 performance optimizations for mobile