--- a/package.json
+++ b/package.json
@@ -39,6 +39,7 @@
     "postcss": "^8.4.31",
     "prettier": "^2.8.0",
     "prettier-plugin-svelte": "^2.10.1",
+    "sharp": "^0.34.3",
     "svelte": "^5.0.0-next.57",
     "svelte-check": "^3.4.3",
     "tailwindcss": "^3.3.5",