… pnpm-lock.yaml
… static/Andres-optimized.webp
… static/Basque.webp
… static/HonestGreens.webp
… static/Isabel-optimized.webp
… static/RecoNombre.webp
… static/VIPS.webp
… static/cinepolis.webp
… static/img/Basic.webp
… static/img/Emp.webp
… static/img/Emp2.webp
… static/img/Gratis2.webp
… static/img/hero-mobile.webp
… static/img/hero.webp