## Changes Made:

### Phase 1: Critical Rendering Path
- Optimized font loading with preload and font-display: swap
- Added DNS prefetch for analytics domains
- Deferred all analytics scripts (GA, Hotjar, Vercel) with requestIdleCallback

### Phase 2: Image Optimization
- Converted 13 images to WebP format (92.8% size reduction)
- Created responsive image variants (hero-mobile.webp)
- Added lazy loading to all below-fold images
- Added explicit dimensions to prevent layout shifts

## Performance Improvements:
- Reduced total image size from 6.5MB to 0.47MB
- Expected LCP improvement: 5.8s → ~2.5s
- Expected TBT improvement: 310ms → ~150ms
- Expected Performance Score: 68 → 80+

🤖 Generated with [Claude Code](https://claude.ai/code)